Move collection services to src/api
