+++ client/src/stores/collectionAttributesStore.ts
@@ -2,8 +2,7 @@ import { defineStore } from "pinia";
 import { computed, del, ref, set } from "vue";
 
 import { DatasetCollectionAttributes } from "@/api";
-
-import { fetchCollectionAttributes } from "./services/datasetCollection.service";
+import { fetchCollectionAttributes } from "@/api/datasetCollections";
 
 export const useCollectionAttributesStore = defineStore("collectionAttributesStore", () => {
     const storedAttributes = ref<{ [key: string]: DatasetCollectionAttributes }>({});
